tender-monitoring-system/requirements.txt
selenium>=4.10.0
beautifulsoup4>=4.9.3
python-telegram-bot>=20.3
webdriver-manager>=4.0.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
async-timeout>=4.0.2
cchardet>=2.1.7
aiodns>=3.0.0
